+++ README.md
@@ -1,0 +1,10 @@
+# Rock, Paper Scissor Game 
+<p> Developed Using Python </p>
+
+## How To Play:
+1. Clone the repo using command `git clone https://github.com/TIgb-ai/Rock-paper-scissor-game-in-py.git `
+2. Open Terminal/PowerShell in your PC.
+3. Navigate to the clone folder.
+4. Run Command `python main.py`
+
+Enjoy!
