+++ artwork.py
@@ -1,0 +1,9 @@
+art="""
+
+    ____             __      ____                           _____      _                     
+   / __ \____  _____/ /__   / __ \____ _____  ___  _____   / ___/_____(_)_____________  _____
+  / /_/ / __ \/ ___/ //_/  / /_/ / __ `/ __ \/ _ \/ ___/   \__ \/ ___/ / ___/ ___/ __ \/ ___/
+ / _, _/ /_/ / /__/ ,<    / ____/ /_/ / /_/ /  __/ /      ___/ / /__/ (__  |__  ) /_/ / /    
+/_/ |_|\____/\___/_/|_|  /_/    \__,_/ .___/\___/_/      /____/\___/_/____/____/\____/_/     
+                                    /_/                                                      
+"""
