--- readme.txt
@@ -1,1 +1,0 @@
-Pls give credits to me if u are sending this to anyone
